feat: ResizableTagTextArea - add support for multiple dependentRef and component refactor
