--- src/Shared/Components/MultipleResizableTextArea/index.ts
@@ -1,2 +1,0 @@
-export * from './MultipleResizableTextArea'
-export * from './types'
