--- src/Shared/Components/MultipleResizableTextArea/types.ts
@@ -1,11 +1,0 @@
-import { DetailedHTMLProps, MutableRefObject, TextareaHTMLAttributes } from 'react'
-
-export interface MultipleResizableTextAreaProps
-    extends DetailedHTMLProps<TextareaHTMLAttributes<HTMLTextAreaElement>, HTMLTextAreaElement> {
-    minHeight?: number
-    maxHeight?: number
-    dataTestId?: string
-    refVar?: MutableRefObject<HTMLTextAreaElement>
-    dependentRefs?: Record<string | number, MutableRefObject<HTMLTextAreaElement>>
-    disableOnBlurResizeToMinHeight?: boolean
-}
