+++ src/Shared/Components/SelectTextArea/SelectTextArea.tsx
@@ -3,6 +3,7 @@ import { SelectInstance } from 'react-select'
 
 import { ReactComponent as ICClearSquare } from '@Icons/ic-clear-square.svg'
 import { useEffectAfterMount } from '@Common/Helper'
+import { ResizableTagTextArea } from '@Common/CustomTagSelector'
 
 import {
     getSelectPickerOptionByValue,
@@ -10,7 +11,6 @@ import {
     SelectPickerOptionType,
     SelectPickerVariantType,
 } from '../SelectPicker'
-import { MultipleResizableTextArea } from '../MultipleResizableTextArea'
 import { SelectTextAreaProps } from './types'
 
 export const SelectTextArea = ({
@@ -21,7 +21,7 @@ export const SelectTextArea = ({
     inputId,
     placeholder,
     refVar,
-    dependentRefs,
+    dependentRef,
     disabled,
     selectPickerProps,
     textAreaProps,
@@ -64,11 +64,11 @@ export const SelectTextArea = ({
     return (
         <div className="select-text-area flexbox dc__align-items-center dc__gap-4 w-100 dc__position-rel">
             {!!Icon && Icon}
-            <MultipleResizableTextArea
+            <ResizableTagTextArea
                 {...textAreaProps}
                 id={inputId}
                 refVar={refVar}
-                dependentRefs={dependentRefs}
+                dependentRef={dependentRef}
                 className={`${textAreaProps?.className || ''} ${disabled ? 'cursor-not-allowed' : ''} ${!value ? 'dc__hide-section' : ''}`}
                 value={value}
                 onChange={onTextAreaChange}
