+++ src/Shared/Components/index.ts
@@ -59,5 +59,4 @@ export * from './Button'
 export * from './InvalidYAMLTippy'
 export * from './ConfirmationModal'
 export * from './DiffViewer'
-export * from './MultipleResizableTextArea'
 export * from './DynamicDataTable'
